Ticket Print Thread Event Times
This commit ensures that if we print thread events on Ticket PDFs, the time is never relative time, even if the agent has their time set to relative. Relative time should only be an option when viewing a Ticket online.

Additionally, we do not need to put anything related to relative time in the ticket-print template.