Issue: Annul Closed Events
This commit fixes an issue where if you were to change a Ticket to the Resolved status and then to Closed, the dashboard statistics would show that you had closed 2 Tickets even though both events were done on the same Ticket.

We should only count one closed event per Ticket.

This was reported in Issue #5018.