issue: Ticket Filter Assignment Event
This addresses 5035 where the ticket Assignment Event show the Ticket
Owner's name or "SYSTEM" as the assigner when tickets are created via Client
Portal, API, or Email Fetching. This updates the event to include "Ticket
Filter" as the assigner if the assignment was triggered by a Ticket Filter
Action. This makes for more accurate events and is very useful for
statistics/debugging. This keeps current functionality of Agents creating
tickets shows the creating Agent's name as the assigner. This also adds a
new parameter to `assignToAgent()` and `assignToTeam()` called `$user`
(defaults to null) that allows us to pass different usernames/user objects
for the events. This makes the two methods versatile so if we add new ways
to assign to Staff/Teams in the future we can pass the new name/object with
no issues.