issue: PHP 7.3 New Agent Set Password
This addresses an issue reported on the Forum where creating a new Agent and
setting a password hangs when using PHP 7.3. This is due to too few
arguments passed to `PasswordPolicy::checkPassword()`. This updates the call
to include a second argument of `null` so the method is satisfied and we can
continue with checking the password.