Issue: Empty Due Date
This commit fixes an issue that allowed Agents to clear the Duedate on a ticket using inline edit. We were saving the value 0000-00-00 00:00:00 to the database which the ticket would then display as 12/1/02 06:09 pm.

Instead, if an Agent clears the Duedate field, we should save the duedate as null and the ticket should continue using the est_duedate.