issue: Complete Thread Var Padding (Outlook)
This addresses issue 5007 where when using the complete thread variable
there is no padding between thread messages in Outlook desktop app and makes
the thread hard to follow. Outlook does not respect some CSS rules that are
accepted in most (if not all) modern email clients. Since we can’t use
normal CSS the workaround is to add a blank table row (with a single space)
so that it forces "padding" between the end of a message and the start of
another.

Before
(image)

After
(image)