feature: Fetch From Mail Folder
This feature adds a textbox in the Email configurations that allows you to
configure a specific Folder to fetch from. If no folder is provided the
system will fetch from the INBOX as usual. This also adds a check so that if
a Folder is given but doesn't exist and can't be created then the system
will error out with a detailed message.