i18n: KnowledgeBase JS
In 5120, the reporting party also mentioned that they receive an error of
`script '/xxx/osTicket/kb/ajax.php' not found or unable to stat`.  This is
due to the AJAX call being inside the KB directory, and `/kb/ajax.php` truly
does not exist. This adds `ROOT_PATH` to the beginning of the AJAX call so
that no matter where it is, it will point to the correct location.