issue: Spaces In Username
This addresses issue 5122 where having a space at the beginning and/or end
of the login Username/Email on the Client Portal the system will return
`Access Denied`. This is due to the system not trimming the rogue spaces.
This adds `trim()` around the Username/Email so that the spaces are removed
and login will be successful.