issue: Allow iFrame Rename
This renames the `Allow iFrames` setting to `Allow System iFrame` as this
better fits the use-case. This also updates the Help Tip information to
include the new title and expands on what you can input (ie. URLs). Lastly,
this renames the validator for the `Allow System iFrame` setting from
`cs-domain` to `cs-url` as it technically validates a comma separated list
of URLs (not just domains).