issue: Attachment Filter
This addresses issue 5123 where fetched Emails do not reject attachments
that are not allowed. This is due to setting a File ID inside of an array
called `$file[]` but not unsetting it when the attachment hits an
FileUploadError. This updates class MailFetch to set the file array back to
empty so that the ID is cleared and the error is added correctly which
rejects the attachment.