php: 5.6 Support
This addresses an issue with v1.14 where `??` (null coalescing operator) in
class Export throws a fatal error when using PHP 5.6. This is an issue
especially for people who cannot upgrade PHP but want to take advantage of
the latest osTicket codebase/features. This updates the null coalescing
operator to an isset ternary operator to keep backwards compatibility with
PHP 5.6.