issue: Redactor Reset Buttons
This addresses an issue where since the Redactor upgrade the Reset buttons
do not work anymore. This is due to the code referencing old plugin/module
code that either no longer exists or has changed names. This updates the
reset button code to reference the new Redactor plugin/module methods.