issue: Attachment Upload Configuration
This addresses an issue where we pull the `first()` Form Field configuration
for attachments instead of determining if we are looking for the Ticket
Attachment Configurations or the Task Attachment Configurations. For
example, if you are creating a Ticket as a User and go to add an attachment
if somehow your Task Details Form comes up first in the configuration lookup
query we will apply the Task Attachment Configs instead of the Ticket
Attachment Configs which is totally incorrect. This updates the lookup for
attachment configurations on web uploads to be specific about what it’s
looking for so it can get the correct configuration. This also adds the
original Form's `type` to SimpleForms so we can access the Form Type when
working with SimpleForms.