support: IE Discontinued
Well folks, the time has finally come to lay IE in it's final resting place.
Twas a long, hard battle with IE but in the end it was defeated and it's
kingdom taken over by it's successor, Edge. Now there is a new king and IE
must be banished from the lands he once ruled, forever. It's sad to see such
a prolific leader fall to the ashes (not lol). Alright, let’s get serious
for a moment. Microsoft is discontinuing IE at the end of January 2020 so
it’s time to start showing a banner that says "osTicket no longer supports
IE." This message is shown on the Installer, the Upgrader, in the Client
Portal, and in the Staff Panel when using IE. This checks the User-Agent
header for IE’s specific header and if detected will show the banner. This
will not show on Edge, Chrome, Firefox, etc. only when using IE. We came to
this conclusion after seeing sites like Github (a microsoft product) showing
"no longer supported" banners for IE (a microsoft product).