issue: mPDF Print Tables
This addresses an issue where printing a Ticket with a Table in one of the
threads will cause the PDF to be malformed. This is due to mPDF config
variable that was changed (most likely by `git checkout` or git conflicts).
This sets the `ignore_table_widths` variable back to true so that tables
will fit the PDF width.