issue: Shared Mailbox Auth
This addresses an issue where using "Servername\user@domain.tld" for mailbox
Login Username fails due to the Shared Mailbox Auth introduced here
(ac9ea5b). We do not check the name to see if the first part of the string
contains an @ symbol, indicating an email address. This updates the Shared
Mailbox Auth check to look for 3 parts, if detected it will use the first
two parts as the `authuser` and the second part as the `username`. This also
checks for 2 parts, if detected it will check the first part for @ symbol
and if one is present will use the first part as the `authuser` and the
second part as the `username`. Otherwise it will use both parts as the
`username`. If there are more than three parts or just one part, the entire
string is used as the `username`.