emoji: Strip From Subject
This addresses an issue reported on the Forum where emojis are not stripped
from the subject line which causes either a blank subject or half-rendered
subject. This adds `Format::strip_emoticons` to TextboxField::parse in class
forms so all Textbox Fields (including Issue Summary) is stripped of Emojis
(since we can't handle them at all). This is also added to
ThreadEntry::create in class thread so that the ThreadEntry Titles have
Emojis stripped as well.