Issue: Help Topic Number Format
This commit fixes an issue where we did not clear the FLAG_CUSTOM_NUMBERS flag when setting the ticket number format back to System Default for Help Topics.

We will also stop clearing saved number_formats so that agents have the option to reenable them without having the reconfigure the format.