Issue: Mass Process Add Users to Orgs
This commit fixes an issue where we could not add multiple users to an organization at a time due to an issue with trying to pass an object to parseJSON rather than a string.

If any of the selected users are already part of a different organization, their organization will NOT be changed.

It also fixes the Organization typeahead to search through parts of an organization name rather than having to type out the whole thing before returning the organization.