Issue: Inline Edit Long Answer
This commit fixes an issue we had with saving an empty long answer field using inline edit. Before, the field would save as an empty line break which would then technically be set and saved, displaying nothing in the field, which prevented agents from doing another inline edit on the long answer field. This is most likely do to the redactor update that makes all the lines double spaced now.