Issue: Saving Ticket CData
This commit fixes an issue where we did not properly save some CData values which caused some fields to not export correctly.

Additionally, when generating the cdata table, we should look at the value on form entry values first rather than looking at the value_id first in the coalesce function.

As a result, the to_php function for several classes of custom fields needs to be adjusted to expect a string value rather than an id.
	- Custom Department Field
	- Custom Assignee Field