feature: Force HTTPS
This feature introduces a checkbox in the System Settings to force HTTPS.
This is for people who either can't or don't know how to configure redirect
rules via the webserver. If the setting is enabled, the system will redirect
all osTicket web traffic to the HTTPS protocol automatically. In addition
this adds an `HTTP_OK` option that disables the HTTPS redirect for local
calls such as cron, pipe, and cli. This also updates the remote piping
scripts (`automail.pl` and `automail.php`) to include support for HTTPS.