Issue: DB Error #1054
Error: Unknown column 'Z1.ticket_id' in 'field list'

This commit fixes an issue with how we approach skipping counts. If a search has a keyword or if a saved queue/search has an invalid criteria saved, we will completely ignore trying to do a count. In saved queue/searches, rather than showing a number or 500, we will display '-' in its place. For the count at the bottom (Showing # - # of about #), we will just display Showing # - #.