Issue: Remove Referrals
This commit fixes an issue where we did not check to see if we needed to remove referrals. For example, if a Ticket is referred to a Department and then the Ticket is transferred to that Department, we should remove the referral.

We will now check/remove referrals as needed for Departments, Agents, and Teams.