Issue: Blank Date Time
This commit fixes an issue where an unset Date Time field would display as 0. As a result, if you were on the all edit page and edited some other field, it would appear as though you edited the date time field as well because it had a value in it. Instead, we need to make sure that the field renders as truly being blank if it isn't set yet to ensure nothing is being saved to it.