queue: Parent/Child "Inherit Columns" Option
This addresses an issue where saving a Parent Queue with "Use standard
columns" Enabled will show the option Enabled but still show the
customizable Columns instead of the Standard Columns. This addresses an
issue where saving a Child Queue with "Inherit columns from the parent
queue" Enabled shows the Standard Columns but does not show the option
Enabled. This adds a check in `inheritColumns()` for Staff Configurations
for the queue and if exists we use the Staff preferences, else we default to
the Queue Flag check. This adds a new method called `staffSettings()` that
returns staff settings for a queue. It takes an optional argument (a
setting's name) that can return a specific setting's value.