queue: AdvancedSearch Use Standard Columns
This addresses an issue where saving a Saved Search with "Use standard
columns" Enabled, shows the Customizable Columns instead of the Standard
Columns and does not show the option Enabled as the INHERIT_COLUMNS flag is
not set.  This adds a check when saving a SavedSearch for the
`inherit-columns` setting and if set, we set the Queue's INHERIT_COLUMNS
Flag.