redactor: Cancel Button
This addresses an issue reported on the Forum where the New Ticket Cancel
buttons do not work in v1.14.1. This is due to the code using older methods
utilized in v1.12. This updates the code to use the new method of accessing
the plugin objects first before accessing the plugin itself.