issue: Update Staff checkPassword()
This addresses an issue where clicking Update in the Set Password modal with
the `Send the agent a password reset email` option checked either shows a
forever spinning Loading screen or instantly kicks back to the modal having
accomplished nothing. This is due to checking `$clean['passwd1']` even when
`welcome_email` (password reset) is set. If `welcome_email` is set/enabled
we should not check the variable as there is no new password being set as we
are sending a Reset Email. This adds a check for `welcome_email` and if
none, then we check the password.