redactor: Double Spacing Optional
This addresses an issue where upgraded redactor does double spacing by
default (which matches most modern editors). A lot of people do not like
this change however so the solution is to make double spacing optional via a
setting. This adds a new Profile setting called Editor Spacing that allows
you to configure Double or Single spacing or the editor. If set to Double,
the text editor will double space when pressing Enter. If set to Single, the
text editor will single space when pressing Enter like the old redactor
version. If a staff does not have this setting configured, it defaults to
Double Spacing.