issue: Require Client Login
This addresses an issue reported on the Forum where having your Client
Registration set to Private and attempting to enable the setting Require
Client Login for Knowledgebase results in an error saying `The knowledge
base cannot be restricted unless client registration is enabled`. When
saving the Knowledgebase Settings and we are updating the Restrict
Knowledgebase setting we check if Client Registration is Enabled using
the `isClientRegistrationEnabled()` method. This method only checks if the
Registration method is `auto` or `open` (Public) and does not check if it
is `closed` (Private). This is an issue in the aforementioned case as Users
can still be registered by Agents so technically registration is
not completely disabled. This removes the check altogether as Admins should
be able to configure this setting without Registration being required. This
also adds a new function called `isClientRegistrationMode()` that can be
used to compare the current Registration Mode to either a string or an array
or modes.