queue: Inherit Columns Option
This is a rewrite of a commit that was accidentally merged. This, in part,
removes all unnecessary code from the accidental commit. This adds a new
`inheritColumns()` method to SavedQueue to use with all queues. Lastly, this
removes `$this->parent_id` from the `$hidden_cols` check so all queues
behave correctly. Most queues will not have a `parent_id` and will fail this
check.