issue: BooleanField Cannot Be Unchecked
This addresses an issue where checkboxes cannot be unchecked once saved as
checked. This was introduced with a fix for another issue where it's
necessary to return the default value for VisibilityConstraints. This is
fine for Holiday Schedule checkboxes but for custom fields this breaks basic
functionality. This updates `getValue()` for CheckboxWidget to only return
the default value if `$data` (source) is not set.