issue: BooleanField Inline Edit Value
This addresses an issue where after setting a checkbox value via Inline Edit
it shows the value as `0` or `1` instead of `Yes` or `No`. You have to
refresh the page in order for the value to appear correctly. This is due to
the raw value being returned instead of it's string equivalent. This adds a
case in `TicketsAjaxAPI::editField()` for `BooleanField` that converts the
raw value to the string value with `toString()`.