issue: New Task Alert
This addresses an issue where the New Task Alert was not functioning as it
should. This adds a new method called `onNewTask()` that runs after a Task
is created and sends out the New Task Alert (if enabled). This will take
into account the Global settings as well as the Department settings for
Alerts and Notices.