vars: Task Subject/Description
This addresses an issue where the `subject` and `description` vars are
missing from `Task::getVar()` so these variables are never replaced in the
Task templates. This adds two cases for `Task::getVar()`; case `description`
that retrieves the Task Description (if any) and case `subject` that
retrieves the Task Subject. This also addresses an issue where the
`description` var is missing from `Task::getVarScope()` so when an Agent
uses the Variable Typeahead the `%{task.description}` option does not show.