issue: Add Remote Collaborator
This addresses an issue where adding a Collaborator that is from a remote
backend such as LDAP fails. This is due to the AJAX URL not being
recognized. The regex is expecting
`^(?P<tid>\d+)/add-collaborator/auth:(?P<bk>\w+):(?P<id>.+)$` but the URL
being called is `XXX/add-collabborator/addcc/auth:XXX` (with the added
`/addcc` part). This updates the regex to account for the `/addcc` part, and
adds `$type` as an argument for `addRemoteCollaborator()`.