issue: Duplicate Tickets in Lookup
This addresses issue 5313 where duplicate tickets appear in the ticket
lookup results; especially when looking for a ticket to merge. This is due
to a discrepancy between both queries looking for hits on the search
criteria. The first query that looks for hits pushed the `number` column
towards the end of the results and the second query has it as the second
column in the results. When we do a `UNION ALL` these values are mapped
incorrectly and it appears as there are 2 results instead of one. This adds
a new method called `lookupByNumber()` that only finds results for number
searches. This also changes the lookup URL for Ticket Merge from `/lookup`
to `/number-lookup` so that we only search for numbers. This also reformats
`lookup()` so that if the search query is a number we will pass to
`lookupByNumber()` and merge the results with `UNION ALL`.