Canned Responses for Dept Roles
This commit fixes a bug where we did not accurately check individual Roles for Departments an Agent could access to determine if they could manage that particular Canned Response.

Now, we check to see if the Agent has a Role that can manage Canned Responses at all to determine if Canned Responses show up in the Knowledge Base.

Additionally, we filter Canned Responses shown based on the Role permissions of the Agent's Departments (Primary and Extended)