issue: Trailing Whitespaces Number Lookups
This addresses issue 5296 where having whitespaces at the beginning or end
of a ticket number lookup (eg. `    643906` or `643906    `) will not show
expected results. This is due to the search query looking for `ticket.number
LIKE '    643906%'` explicitly which will not find results. This updates the
Number Lookup method to trim all white spaces from number criteria so the
`LIKE` comparison operator works as expected and finds the appropriate
results.