issue: Multiple Choice Export
This addresses an issue where saving multiple choices for a ChoiceField does
not show all values in Queue Columns nor Ticket Exports. This is due to the
`to_php()` method exploding on `,` while the keys are saved like `key1,
key2` (with comma then space). This updates the `to_php()` method to explode
on `, ` (comma and space) so we get the proper keys for the value lookups.