install: Task Title/Description Field Flags
This addresses an issue where the Task Title field allows any Admin to
remove the View and Required perms for Agents even though this field should
be Viewable and Required. This adds the `FLAG_MASK_REQUIRE` and
`FLAG_MASK_VIEW` flags to the form install file to ensure Admins cannot
change this behavior for both perms. This also addresses an issue where the
Task Description field allows any Admin to remove the View perms for Agents
even though this field should always be Viewable. This adds the
`FLAG_MASK_VIEW` flag to the form install file to ensure Admins cannot
change this behavior.