issue: ThreadEntryField Help Text
This addresses an issue introduced with the previous commit where since we
allow Formatting to be saved for Help Text now, the ThreadEntryFields do not
render the Help Text correctly. Only ThreadEntryFields display the Help Text
within the `placeholder` attribute. Since the content is placed in an
attribute, it cannot be rendered as HTML. This adds `Format::striptags()` to
`ThreadEntryField::getWidget()` so that tags are stripped from the Help Text
before rendering.