i18n: TextareaField Placeholder
This addresses an issue where TextareaFields do not show translated
placeholders. This is due to using the placeholder from the default
configuration instead of using the local placeholder. This updates the
`TextareaWidget::render()` method to get the local placeholder instead of
using the default from config. This also passes `$config['placeholder']` as
a default in case there is no translation available.