issue: Information Field Help Text Decode
This addresses an issue where formatting Help Text for Information fields
displays the content as plain-text instead of rendering the HTML. This is
due to `htmlchars()` which uses `htmlspecialchars()` to encode the content.
This updates `Format::htmlchars()` to `Format::display()` so we can render
the content as HTML rather than plaintext.