oops: is_numeric Soft Fail
This addresses an issue introduced with 9e21df where `is_numeric()`
validation fails even when inserting a legitimate number. In the
aforementioned commit, we added the ability to return error messages from
the source validator method. In doing so, we changed `call_user_func()` to
`call_user_func_array()` which now passes two arguments to the validation
method. This broke `is_numeric()` validation as this method does not accept
more than one argument. This adds a new method to class Validator called
`is_numeric()` that accepts two arguments so it does not fail. In addition,
this updates the remaining validators that do not currently accept two
arguments.