issue: Department Parent
This addresses issue 5428 where you can set a Department's Parent to itself,
which is not supposed to happen. This adds a check to `Department::update()`
that checks if the current Department is in the Parent's path. If so, we
will refuse to save changes and throw an error message stating `Parent
Department cannot contain the current Department`. This also fixes a small
issue where we were not using `$errors['pid']`, instead we were using
`$warn`. This updates `$warn` to `$errors['pid']` so we can use
`$errors['pid']` if set elsewhere.